% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimized_scatter.R
\name{apply_kernel_optimized}
\alias{apply_kernel_optimized}
\title{apply_kernel_optimized}
\usage{
apply_kernel_optimized(
  raw_input,
  out_size,
  filter = "circle",
  mask = default_kernel(filter, radius, sigma),
  radius = 2,
  sigma = radius/2
)
}
\arguments{
\item{raw_input}{Raw vector containing RGBA bitmap data.}

\item{out_size}{2-element vector size of the bitmap.}

\item{filter}{Use the pre-defined filter, either \code{circle}, \code{square}, \code{gauss}. Defaults to \code{circle}.}

\item{mask}{Custom kernel used for blurring, overrides \code{filter}. Must be a square matrix of odd size.}

\item{radius}{Radius of the kernel (counted without the "middle" pixel"), defaults to 2. The generated kernel matrix will be a square with (2*radius+1) pixels on each side.}

\item{sigma}{Radius of the Gaussian function selected by \code{filter}, defaults to \code{radius/2}.}
}
\value{
A raw vector containing the blurred RGBA bitmap data.
}
\description{
Optimized version of apply_kernel_rgbwt that uses raw uint8_t arrays for faster performance.
}
