% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{Color103}
\alias{Color103}
\title{Predefined color palettes for flow cytometry populations}
\format{
An object of class \code{list} of length 2.
}
\usage{
Color103
}
\description{
Predefined color palettes for flow cytometry populations
}
\keyword{datasets}
