% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{Color103}
\alias{Color103}
\title{Color palette with 103 distinct colors}
\format{
An object of class \code{character} of length 102.
}
\usage{
Color103
}
\description{
Color palette with 103 distinct colors
}
\keyword{datasets}
