% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_kernel.R
\name{default_kernel}
\alias{default_kernel}
\title{default_kernel}
\usage{
default_kernel(filter, radius, sigma)
}
\description{
Create a nice defaulted kernel for \verb{apply_kernel_*} functions.
Parameters are documented in the exported functions.
}
