% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f.axis_label}
\alias{f.axis_label}
\title{Add axis labels to flow cytometry plots}
\usage{
f.axis_label(labels, padding = 1.3, font = 2, font.size = 0.8)
}
\arguments{
\item{labels}{Axis labels}

\item{padding}{Padding for labels}

\item{font}{Font style}

\item{font.size}{Font size}
}
\description{
Add axis labels to flow cytometry plots
}
