% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f.axis_ticks}
\alias{f.axis_ticks}
\title{Draw axis ticks for flow cytometry plots}
\usage{
f.axis_ticks(
  p_chnames,
  major_tick_loc,
  minor_tick_loc,
  ranges,
  shownumbers = FALSE
)
}
\arguments{
\item{p_chnames}{Channel names}

\item{major_tick_loc}{Major tick locations}

\item{minor_tick_loc}{Minor tick locations}

\item{ranges}{Plot ranges}

\item{shownumbers}{Whether to show tick labels}
}
\description{
Draw axis ticks for flow cytometry plots
}
