% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f.parseChannels}
\alias{f.parseChannels}
\title{Parse channel names or indices from a flow cytometry object}
\usage{
f.parseChannels(obj, channels)
}
\arguments{
\item{obj}{A flow frame object}

\item{channels}{Channel names or indices}
}
\value{
Numeric indices of the channels
}
\description{
Parse channel names or indices from a flow cytometry object
}
