% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f.plotGate}
\alias{f.plotGate}
\title{Plot gates on flow cytometry plots}
\usage{
f.plotGate(x, xlim, ylim, log_ssc)
}
\arguments{
\item{x}{Gate definitions}

\item{xlim}{X-axis limits}

\item{ylim}{Y-axis limits}

\item{log_ssc}{Whether SSC is log-transformed}
}
\description{
Plot gates on flow cytometry plots
}
