% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f.transFscSscData}
\alias{f.transFscSscData}
\title{Transform FSC/SSC data}
\usage{
f.transFscSscData(x, method)
}
\arguments{
\item{x}{Data vector}

\item{method}{Transformation method}
}
\value{
Transformed data
}
\description{
Transform FSC/SSC data
}
