% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f.transVectData}
\alias{f.transVectData}
\title{Transform flow cytometry data}
\usage{
f.transVectData(x, method, parameters = transform_parameters)
}
\arguments{
\item{x}{Data vector}

\item{method}{Transformation method}

\item{parameters}{Transformation parameters}
}
\value{
Transformed data
}
\description{
Transform flow cytometry data
}
