% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowPlot.R
\name{flowPlot}
\alias{flowPlot}
\title{Create a 2D density plot for flow cytometry data}
\usage{
flowPlot(
  obj,
  channels,
  title = "",
  title.font.size = 1.2,
  res = 256,
  color = "#00000010",
  xlim = NA,
  ylim = NA,
  plot_ceiling = NA,
  density.overlay = TRUE,
  dens.color = "red",
  dens.alpha = 0.5
)
}
\arguments{
\item{obj}{A flow frame, matrix, or data frame containing flow cytometry data}

\item{channels}{Channels to plot (column indices or names)}

\item{title}{Plot title}

\item{title.font.size}{Font size for the title}

\item{res}{Resolution of the plot}

\item{color}{Color for the points}

\item{xlim}{X-axis limits}

\item{ylim}{Y-axis limits}

\item{plot_ceiling}{Maximum value for plotting}

\item{density.overlay}{Whether to overlay density plots}

\item{dens.color}{Color for density overlay}

\item{dens.alpha}{Alpha transparency for density overlay}
}
\value{
A plot of the flow cytometry data
}
\description{
A simple implementation of scattermore functionality for flow cytometry data visualization.
}
