% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{major_tick}
\alias{major_tick}
\title{Major tick marks for log scale}
\format{
An object of class \code{numeric} of length 12.
}
\usage{
major_tick
}
\description{
Major tick marks for log scale
}
\keyword{datasets}
