% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{major_tick_linear}
\alias{major_tick_linear}
\title{Major tick marks for linear scale}
\format{
An object of class \code{numeric} of length 6.
}
\usage{
major_tick_linear
}
\description{
Major tick marks for linear scale
}
\keyword{datasets}
