% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{minor_tick}
\alias{minor_tick}
\title{Minor tick marks for log scale}
\format{
An object of class \code{numeric} of length 73.
}
\usage{
minor_tick
}
\description{
Minor tick marks for log scale
}
\keyword{datasets}
