% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{minor_tick_linear}
\alias{minor_tick_linear}
\title{Minor tick marks for linear scale}
\format{
An object of class \code{numeric} of length 23.
}
\usage{
minor_tick_linear
}
\description{
Minor tick marks for linear scale
}
\keyword{datasets}
