% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{number_to_color}
\alias{number_to_color}
\title{Map numeric values to color gradients}
\usage{
number_to_color(
  numbers,
  palette = "magma",
  lower_percentile = 0.01,
  upper_percentile = 0.99,
  breaks = 256
)
}
\arguments{
\item{numbers}{Numeric values}

\item{palette}{Color palette}

\item{lower_percentile}{Lower percentile for normalization}

\item{upper_percentile}{Upper percentile for normalization}

\item{breaks}{Number of color bins}
}
\value{
Character vector of hex color codes
}
\description{
Map numeric values to color gradients
}
