% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFlowFrame_optimized.R
\name{PlotFlowFrame_optimized}
\alias{PlotFlowFrame_optimized}
\title{Advanced flow cytometry plotting function (Optimized Version)}
\usage{
PlotFlowFrame_optimized(
  obj,
  channels,
  colormap_marker = NA,
  pop_idex.list = NA,
  plotting_background = FALSE,
  title,
  title.font.size,
  comp = FALSE,
  spillovermatix = NA,
  min.xy,
  max.xy,
  lowerlimit.quantile = 3e-04,
  showticks = TRUE,
  shownumbers = FALSE,
  showlabels = TRUE,
  plot.margins = c(2, 2, 2, 2),
  res = NA,
  dens_res,
  population_colors = Color103,
  auto_assign_color = TRUE,
  background_color = "#525252",
  rgba_blending = TRUE,
  interpolate = FALSE,
  rare_on_top = TRUE,
  highlight_pops = NA,
  transformation = "logicle",
  tr.parameters = NA,
  print_estimated_parameters = FALSE,
  default_b = 400,
  log_fsc = FALSE,
  log_ssc = FALSE,
  label_size = 0.8,
  color_pops = TRUE,
  gates,
  old_method = TRUE,
  density.overlay = FALSE,
  max_points = 5e+05,
  use_raw_arrays = TRUE
)
}
\arguments{
\item{obj}{A flow frame, GatingSet, or FlowSet object}

\item{channels}{Channels to be plotted}

\item{colormap_marker}{Channel to use for color mapping}

\item{pop_idex.list}{List of cell indices for different populations}

\item{plotting_background}{Whether to plot background data}

\item{title}{Plot title}

\item{title.font.size}{Font size for title}

\item{comp}{Whether to apply compensation}

\item{spillovermatix}{Spillover matrix for compensation}

\item{min.xy}{Minimum values for x and y axes}

\item{max.xy}{Maximum values for x and y axes}

\item{lowerlimit.quantile}{Lower limit quantile for trimming outliers}

\item{showticks}{Whether to show axis ticks}

\item{shownumbers}{Whether to show axis numbers}

\item{showlabels}{Whether to show axis labels}

\item{plot.margins}{plotting margins}

\item{res}{Resolution for plotting}

\item{dens_res}{Resolution for density estimation}

\item{population_colors}{Colors for populations}

\item{auto_assign_color}{Whether to auto-assign colors}

\item{background_color}{Color for background}

\item{rgba_blending}{Whether to use RGBA blending}

\item{interpolate}{Whether to interpolate plot}

\item{rare_on_top}{Whether to plot rare populations on top}

\item{highlight_pops}{Populations to highlight}

\item{transformation}{Transformation type (e.g., arcsinh, logicle)}

\item{tr.parameters}{Parameters for transformations}

\item{print_estimated_parameters}{Whether to print estimated parameters}

\item{default_b}{Default value for arcsinh transformation}

\item{log_fsc}{Whether to log-transform FSC channels}

\item{log_ssc}{Whether to log-transform SSC channels}

\item{label_size}{Size for axis labels}

\item{color_pops}{Whether to enable population coloring}

\item{gates}{Gates to be drawn}

\item{old_method}{Whether to use older rendering method}

\item{density.overlay}{Whether to overlay density plots}

\item{max_points}{Maximum number of points to plot (for downsampling)}

\item{use_raw_arrays}{Whether to use raw uint8_t arrays for faster performance. Default is TRUE.}
}
\value{
A plot of the flow cytometry data
}
\description{
Create advanced visualizations of flow cytometry data with multiple options for transformations,
population coloring, and density overlays. This is an optimized version that uses raw uint8_t arrays
for faster performance.
}
