% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimized_scatter.R
\name{raw_to_raster}
\alias{raw_to_raster}
\title{raw_to_raster}
\usage{
raw_to_raster(raw_data, out_size)
}
\arguments{
\item{raw_data}{Raw vector containing RGBA bitmap data.}

\item{out_size}{2-element vector size of the bitmap.}
}
\value{
A raster object.
}
\description{
Convert a raw RGBA bitmap to a raster object for plotting.
}
