% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scientific}
\alias{scientific}
\title{Format numbers as scientific expressions}
\usage{
scientific(x)
}
\arguments{
\item{x}{Numeric values to format}
}
\value{
Formatted scientific expressions
}
\description{
Format numbers as scientific expressions
}
