% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFlowFrame.R
\name{summary,flowCore::transformList-method}
\alias{summary,flowCore::transformList-method}
\title{Advanced flow cytometry plotting function}
\usage{
\S4method{summary}{flowCore::transformList}(object, ...)
}
\description{
Create advanced visualizations of flow cytometry data with multiple options for transformations,
population coloring, and density overlays.
}
